addressed the bitshifting signed integers problem
This problem originally brought up here: https://github.com/lcm-proj/lcm/pull/31

Only added the changes from the single commit that was addressing the actual
shifting errors:
https://github.com/lcm-proj/lcm/commit/63963c7d58cd27c351274b38a56343c606ae21af

+++ zcm/util/ioutils.h
@@ -21,7 +21,7 @@ static inline int fwrite32(FILE *f, int32_t v)
 
 static inline int fwrite64(FILE *f, int64_t v64)
 {
-    int32_t v = v64>>32;
+    int32_t v = ((uint64_t)v64)>>32;
     if (0 != fwrite32(f, v)) return -1;
     v = v64 & 0xffffffff;
     return fwrite32(f, v);
@@ -49,7 +49,7 @@ static inline int fread64(FILE *f, int64_t *v64)
     if (fread32(f, &v2))
         return -1;
 
-    *v64 =     (((int64_t) v1)<<32) | (((int64_t) v2)&0xffffffff);
+    *v64 = (int64_t)(((uint64_t)v1)<<32) | (((int64_t)v2)&0xffffffff);
 
     return 0;
 }
